Copy fix from be3ffc72975d01a5b39722a4ff55269eb69afe93

+++ weblab/experiments/signals.py
@@ -6,5 +6,8 @@ def experiment_version_deleted(sender, instance, **kwargs):
     Signal callback when an experiment version is about to be deleted.
 
     Ensure the experiment data directory is also deleted.
+
+    If the directory doesn't exist yet (because no results received) this is a no-op.
     """
-    rmtree(str(instance.abs_path))
+    if instance.abs_path.is_dir():
+        rmtree(str(instance.abs_path))
